rewrite dmg build script to avoid using Finder which times out when running headless

+++ cmake/packaging/build_macos_dmg.sh
@@ -4,35 +4,41 @@ APP_NAME=lammps-gui
 VERSION="$1"
 LAMMPS_GUI_APP="$2"
 BUILD_DIR="${PWD}"
+PACKAGING_DIR="$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)"
+STAGE_DIR="${BUILD_DIR}/dmg-staging"
+DMG_FILE="LAMMPS-macOS-multiarch-GUI-${VERSION}.dmg"
+
+# install/upgrade dmgbuild helper script
+python3 -m pip install --upgrade --user pip
+python3 -m pip install --upgrade --user dmgbuild
+
+if ! python3 -c 'import dmgbuild' > /dev/null 2>&1
+then
+    echo "ERROR: dmgbuild is required. Install with: python3 -m pip install --user dmgbuild"
+    exit 1
+fi
+
 rm -rv ${APP_NAME}.app
 mv -v ${LAMMPS_GUI_APP} .
 
 echo "Delete old files, if they exist"
 rm -f ${APP_NAME}.dmg ${APP_NAME}-rw.dmg LAMMPS-macOS-multiarch-GUI-*.dmg
+rm -rf "${STAGE_DIR}"
 
 echo "Sign dynamic LAMMPS library and LAMMPS-GUI"
 codesign --force -s - ${BUILD_DIR}/liblammps.0.dylib
 codesign --force -s - ${BUILD_DIR}/lammps-gui.app/Contents/MacOS/lammps-gui
 codesign --force -s - ${BUILD_DIR}/lammps-gui.app/Contents/Frameworks/liblammps.0.dylib
 
-echo "Create initial dmg file with macdeployqt"
-macdeployqt ${APP_NAME}.app -dmg
-echo "Create writable dmg file"
-hdiutil convert ${APP_NAME}.dmg -format UDRW -o ${APP_NAME}-rw.dmg
+echo "Bundle Qt frameworks and plugins with macdeployqt"
+macdeployqt ${APP_NAME}.app
 
-echo "Mount writeable DMG file in read-write mode. Keep track of device and volume names"
-DEVICE=$(hdiutil attach -readwrite -noverify ${APP_NAME}-rw.dmg | grep '^/dev/' | sed 1q | awk '{print $1}')
-VOLUME=$(df | grep ${DEVICE} | sed -e 's/^.*\(\/Volumes\/\)/\1/')
-sleep 2
-
-echo "Create link to Application folder and move README and background image files"
-
-pushd "${VOLUME}"
-ln -s /Applications .
-mv ${APP_NAME}.app/Contents/Resources/README.txt .
-mkdir  .background
-mv ${APP_NAME}.app/Contents/Resources/LAMMPS_DMG_Background.png .background/background.png
-mv ${APP_NAME}.app LAMMPS-GUI.app
+echo "Stage a copy of the app bundle plus README and background image"
+mkdir -p "${STAGE_DIR}"
+ditto ${APP_NAME}.app "${STAGE_DIR}/LAMMPS-GUI.app"
+pushd "${STAGE_DIR}"
+mv LAMMPS-GUI.app/Contents/Resources/README.txt .
+mv LAMMPS-GUI.app/Contents/Resources/LAMMPS_DMG_Background.png background.png
 cd LAMMPS-GUI.app/Contents
 
 echo "Update rpath for LAMMPS and LAMMPS-GUI to link to bundled liblammps.0.dylib"
@@ -41,8 +47,10 @@ install_name_tool -add_rpath '@executable_path/../Frameworks' bin/lmp
 
 echo "Codesign bundled plugins"
 codesign --force -s - PlugIns/*/*.dylib
+
 echo "Codesign bundled frameworks"
 codesign --force -s - Frameworks/Qt*.framework/Versions/A/Qt*
+
 echo "Codesign bundled executables"
 for s in bin/*
 do \
@@ -67,78 +75,20 @@ fi
 rm icon.rsrc
 popd
 
-echo 'Tell the Finder to resize the window, set the background,'
-echo 'change the icon size, place the icons in the right position, etc.'
-echo '
-    tell application "Finder"
-    tell disk "'${APP_NAME}'"
-
-      -- wait for the image to finish mounting
-      set open_attempts to 0
-      repeat while open_attempts < 4
-        try
-          open
-            delay 1
-            set open_attempts to 5
-          close
-        on error errStr number errorNumber
-          set open_attempts to open_attempts + 1
-          delay 10
-        end try
-      end repeat
-      delay 5
-
-      -- open the image the first time and save a .DS_Store
-      -- just the background and icon setup
-      open
-        set current view of container window to icon view
-        set theViewOptions to the icon view options of container window
-        set background picture of theViewOptions to file ".background:background.png"
-        set arrangement of theViewOptions to not arranged
-        set icon size of theViewOptions to 64
-        delay 5
-      close
-
-      -- next set up the position of the app and Applications symlink
-      -- plus hide all window decorations
-      open
-        update without registering applications
-        tell container window
-          set sidebar width to 0
-          set statusbar visible to false
-          set toolbar visible to false
-          set the bounds to { 100, 40, 868, 640 }
-          set position of item "'LAMMPS-GUI'.app" to { 190, 216 }
-          set position of item "Applications" to { 576, 216 }
-          set position of item "README.txt" to { 190, 400 }
-        end tell
-        update without registering applications
-        delay 5
-      close
-
-      -- one last open and close to check the results
-      open
-        delay 5
-      close
-    end tell
-    delay 1
-  end tell
-' | osascript
-
-sync
-
-echo "Unmount modified disk image and convert to compressed read-only image"
-hdiutil detach "${DEVICE}"
-hdiutil convert "${APP_NAME}-rw.dmg" -format UDZO -o "LAMMPS-macOS-multiarch-GUI-${VERSION}.dmg"
+echo "Create compressed disk image with Finder layout using dmgbuild (no AppleScript)"
+python3 -m dmgbuild -s "${PACKAGING_DIR}/dmg_settings.py" \
+    -D app="${STAGE_DIR}/LAMMPS-GUI.app" \
+    -D readme="${STAGE_DIR}/README.txt" \
+    -D background="${STAGE_DIR}/background.png" \
+    "${APP_NAME}" "${DMG_FILE}"
 
 echo "Attach icon to .dmg file"
 echo "read 'icns' (-16455) \"${APP_NAME}.app/Contents/Resources/lammps.icns\";" > icon.rsrc
-Rez -a icon.rsrc -o LAMMPS-macOS-multiarch-GUI-${VERSION}.dmg
-SetFile -a C LAMMPS-macOS-multiarch-GUI-${VERSION}.dmg
+Rez -a icon.rsrc -o ${DMG_FILE}
+SetFile -a C ${DMG_FILE}
 rm icon.rsrc
 
-echo "Delete temporary disk images"
-rm -f "${APP_NAME}-rw.dmg"
-rm -f "${APP_NAME}.dmg"
+echo "Delete staging directory"
+rm -rf "${STAGE_DIR}"
 
 exit 0
